Refactor act 3 actors into `act3actors.cpp`
