+++ CMakeLists.txt
@@ -295,9 +295,6 @@ function(add_lego_libraries NAME)
     LEGO1/lego/legoomni/src/actors/act2genactor.cpp
     LEGO1/lego/legoomni/src/actors/act3actors.cpp
     LEGO1/lego/legoomni/src/actors/act3ammo.cpp
-    LEGO1/lego/legoomni/src/actors/act3brickster.cpp
-    LEGO1/lego/legoomni/src/actors/act3cop.cpp
-    LEGO1/lego/legoomni/src/actors/act3shark.cpp
     LEGO1/lego/legoomni/src/actors/ambulance.cpp
     LEGO1/lego/legoomni/src/actors/bike.cpp
     LEGO1/lego/legoomni/src/actors/buildingentity.cpp
