--- LEGO1/lego/legoomni/src/actors/act3brickster.cpp
@@ -1,55 +1,0 @@
-#include "act3brickster.h"
-
-DECOMP_SIZE_ASSERT(Act3Brickster, 0x1b4)
-
-// STUB: LEGO1 0x10040e10
-Act3Brickster::Act3Brickster()
-{
-	// TODO
-}
-
-// FUNCTION: LEGO1 0x10040f20
-Act3Brickster::~Act3Brickster()
-{
-	// TODO
-}
-
-// STUB: LEGO1 0x10040ff0
-void Act3Brickster::ParseAction(char* p_extra)
-{
-	// TODO
-}
-
-// STUB: LEGO1 0x10041050
-void Act3Brickster::VTable0x70(float p_time)
-{
-	// TODO
-}
-
-// STUB: LEGO1 0x100416b0
-MxResult Act3Brickster::VTable0x94(LegoPathActor*, MxBool)
-{
-	// TODO
-	return SUCCESS;
-}
-
-// STUB: LEGO1 0x100417c0
-// STUB: BETA10 0x1001a407
-MxResult Act3Brickster::FUN_100417c0()
-{
-	// TODO
-	return SUCCESS;
-}
-
-// STUB: LEGO1 0x10042990
-void Act3Brickster::SwitchBoundary(LegoPathBoundary*& p_boundary, LegoUnknown100db7f4*& p_edge, float& p_unk0xe4)
-{
-	// TODO
-}
-
-// STUB: LEGO1 0x100429d0
-MxResult Act3Brickster::VTable0x9c()
-{
-	// TODO
-	return SUCCESS;
-}
