+++ LEGO1/lego/legoomni/src/common/legoobjectfactory.cpp
@@ -6,9 +6,6 @@
 #include "act2policestation.h"
 #include "act3.h"
 #include "act3actors.h"
-#include "act3brickster.h"
-#include "act3cop.h"
-#include "act3shark.h"
 #include "ambulance.h"
 #include "bike.h"
 #include "buildings.h"
